Create BREAST CANCER WISCONSIN PREDICTION
Breast cancer is the most common cancer among women in the world. It accounts for 25% of all cancer cases and affected over 2.1 Million people in 2015 alone. It starts when cells in the breast begin to grow out of control. These cells usually form tumors that can be seen via X-ray or felt as lumps in the breast area. Early detection is the best way to increase the chance of treatment and survivability. Data mining has become a popular tool for knowledge discovery which shows good results in marketing, social science, finance, and medicine.

The key challenge against it’s detection is how to classify tumors into malignant (cancerous) or benign(non cancerous).